+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/MessageFooter.kt
@@ -81,7 +81,10 @@ public fun MessageFooter(messageItem: MessageItemState) {
                 )
             }
 
-            Timestamp(date = message.updatedAt ?: message.createdAt)
+            val date = message.updatedAt ?: message.createdAt ?: message.createdLocallyAt
+            if (date != null) {
+                Timestamp(date = date)
+            }
         }
     }
 }