+++ stream-chat-android-pushprovider-firebase/src/main/java/io/getstream/chat/android/pushprovider/firebase/FirebasePushDeviceGenerator.kt
@@ -28,8 +28,9 @@ import io.getstream.chat.android.client.notifications.handler.PushDeviceGenerato
 /**
  * Generator responsible for providing information needed to register Firebase push notifications provider
  */
-public class FirebasePushDeviceGenerator(private val firebaseMessaging: FirebaseMessaging = FirebaseMessaging.getInstance()) :
-    PushDeviceGenerator {
+public class FirebasePushDeviceGenerator(
+    private val firebaseMessaging: FirebaseMessaging = FirebaseMessaging.getInstance(),
+) : PushDeviceGenerator {
     private val logger = ChatLogger.get("ChatNotifications")
 
     override fun isValidForThisDevice(context: Context): Boolean =